move blob operations to blobs.ex
